+++ apps/examples/src/yjs/YjsExample.tsx
@@ -3,8 +3,15 @@ import '@tldraw/tldraw/editor.css'
 import '@tldraw/tldraw/ui.css'
 import { useYjsStore } from './useYjsStore'
 
+const HOST_URL =
+	process.env.NODE_ENV === 'development' ? 'ws://localhost:1234' : 'wss://demos.yjs.dev'
+
 export default function YjsExample() {
-	const store = useYjsStore()
+	const store = useYjsStore({
+		roomId: 'example',
+		hostUrl: HOST_URL,
+		version: 1,
+	})
 
 	return (
 		<div className="tldraw__editor">
